tenant page components refactoring
