+++ packages/admin/src/tenant/components/index.ts
@@ -2,6 +2,7 @@ export * from './apiKey'
 export * from './idp'
 export * from './member'
 export * from './otp'
+export * from './pages'
 export * from './password'
 export * from './person'
 export * from './project'
