+++ packages/admin/src/tenant/components/pages/errors.tsx
@@ -1,0 +1,3 @@
+export const createNotInProjectError = (): Error => {
+	return new Error('This component can be only used in a project context. Wrap it in <ProjectSlugProvider/>.')
+}
