+++ packages/admin/src/tenant/components/pages/index.ts
@@ -1,0 +1,3 @@
+export * from './InviteUserPage'
+export * from './EditUserPage'
+export * from './UserListPage'
